requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
cycler==0.10.0
Django==3.0.2
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
jieba==0.40
kiwisolver==1.1.0
lxml==4.4.2
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.9.0
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
requests==2.22.0
shodan==1.21.2
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wordcloud==1.6.0
XlsxWriter==1.2.7
yapf==0.29.0
